refactor(ProductObserverData):moved catalog category product logic to a service for using it globally
